pyproject.toml
[tool.poetry]
name = "grape-pi"
version = "1.0.1"
description = "Using graph neural networks to enhance protein identification using protein interaction networks."
authors = ["Chunhui Gu <fduguchunhui@gmail.com>"]
readme = "README.md"
packages = [
    { include = "src"},
    { include = "grape_pi.py"},
    { include = 'main.py'}
]
include=[
    "configs/protein",
    "README.md",
]

[tool.poetry.dependencies]
python = ">=3.11.0, <=3.12.0"
torch = "2.4.0"
numpy = "^1.26.4"
pandas = "^2.2.2"
yacs = "^0.1.8"
hydra-core = "^1.3.2"
chardet = "^5.2.0"
pytorch-lightning = "^2.2.5"
h5py = "^3.11.0"
ogb = "^1.3.6"
overrides = "^7.7.0"
psutil = "^5.9.8"
pyparsing = "^3.1.2"
tensorboardx = "^2.6.2.2"
torch-geometric = "^2.6.1"
seaborn = "0.12.0"
jupyter = {version="^1.1.1", optional=true}
gseapy = {version="^1.1.3", optional=true}
matplotlib = {version="^3.9.0", optional=true}
qtconsole = {version = "^5.6.1", optional = true}
matplotlib-venn = {version = "^1.1.1", optional = true}


[[tool.poetry.source]]
name = "PyPI"
priority = "primary"

[[tool.poetry.source]]
name = "pyg"
url = "https://data.pyg.org/whl/torch-2.4.0+cpu.html"

[tool.poetry.scripts]
grapepi = "grape_pi:main"
grapepi_train = "main:main"

[tool.poetry.group.dev.dependencies]
flake8 = "^7.1.1"

[tool.poetry.group.test.dependencies]
pytest = "^8.3.3"

[tool.poetry.group.additional.dependencies]
torch-scatter = [
    {markers="sys_platform == 'darwin' and python_version >= '3.11' and python_version < '3.12' ", url =  "https://data.pyg.org/whl/torch-2.4.0+cpu/torch_scatter-2.1.2-cp311-cp311-macosx_10_9_universal2.whl"},
    {markers="sys_platform == 'darwin' and python_version >= '3.12' and python_version < '3.13'", url =  "https://data.pyg.org/whl/torch-2.4.0+cpu/torch_scatter-2.1.2-cp312-cp312-macosx_10_9_universal2.whl"},
    {version='2.1.2+pt24cpu', markers="sys_platform != 'darwin'", source='pyg'}
]

torch-sparse = [
    {markers="sys_platform == 'darwin' and python_version >= '3.11' and python_version < '3.12' ", url =  "https://data.pyg.org/whl/torch-2.4.0+cpu/torch_sparse-0.6.18-cp311-cp311-macosx_11_0_universal2.whl"},
    {markers="sys_platform == 'darwin' and python_version >= '3.12' and python_version < '3.13' ", url =  "https://data.pyg.org/whl/torch-2.4.0+cpu/torch_sparse-0.6.18-cp312-cp312-macosx_11_0_universal2.whl"},
    {version='*', markers="sys_platform != 'darwin'", source='pyg'},
]

torch-cluster = [
    {markers="sys_platform == 'darwin' and python_version >= '3.11' and python_version < '3.12' ", url =  "https://data.pyg.org/whl/torch-2.4.0+cpu/torch_cluster-1.6.3-cp311-cp311-macosx_10_9_universal2.whl"},
    {markers="sys_platform == 'darwin' and python_version >= '3.12' and python_version < '3.13' ", url =  "https://data.pyg.org/whl/torch-2.4.0+cpu/torch_cluster-1.6.3-cp312-cp312-macosx_10_9_universal2.whl"},
    {version='*', markers="sys_platform != 'darwin'", source='pyg'}
]

torch-spline_conv = [
    {markers="sys_platform == 'darwin' and python_version >= '3.11' and python_version < '3.12' ", url =  "https://data.pyg.org/whl/torch-2.4.0+cpu/torch_spline_conv-1.2.2-cp311-cp311-macosx_10_9_universal2.whl"},
    {markers="sys_platform == 'darwin' and python_version >= '3.12' and python_version < '3.13' ", url =  "https://data.pyg.org/whl/torch-2.4.0+cpu/torch_spline_conv-1.2.2-cp312-cp312-macosx_10_9_universal2.whl"},
    {version='*', markers="sys_platform != 'darwin'", source='pyg'}
]

[tool.poetry.extras]
notebooks = ["seaborn", "jupyter", "matplotlib", "qtconsole", "gseapy", 'matplotlib-venn']

[build-system]
requires = ["poetry-core>=1.8.3"]
build-backend = "poetry.core.masonry.api"